refactor(scheduler): extract engine to ovp-scheduler crate; codex round-4 fixes
- Move the reusable engine (Cadence, Registry, State, is_due, plan_tick,
  job_shell_command, JobRunner) into a new ovp-scheduler library crate so a
  future desktop app shares one engine (AGENTS.md: ovp-cli stays thin) (P2).
  ovp-cli/scheduler.rs is now a thin adapter (process runner, clock, lock,
  output, CliError mapping).
- run_tick persists state after EACH job via plan_tick, so an interrupted tick
  never reruns a completed job on the next tick (P2).
- install aborts if the legacy com.ovp2.daily unit is still loaded after a
  failed stop, instead of orphaning a live job whose file it deleted (P1).
- crystal-themes publishes themes.json atomically (temp+rename) so crystal-synth
  can't read a truncated projection (P2).
- install holds the dispatch lock across registry reconcile + state seeding, so
  a concurrent enable/disable/desktop edit isn't lost (P2).
