+++ crates/ovp-scheduler/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "ovp-scheduler"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Reusable job-scheduling engine: cadence math, job registry + state, due calculation, and dispatch policy (shared by the CLI and a future desktop app)"
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
+chrono = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
